requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
email-validator
Flask-WTF
Flask-Login
python-dotenv
Werkzeug
WTForms-SQLAlchemy

python-magic
Flask-Limiter

Flask-Mail